Generate matches / Displaing matches in Touranament Upcoming matches

+++ app/src/main/java/com/example/tournamentapp/database/match/SingleMatch.kt
@@ -27,5 +27,5 @@ data class SingleMatch(
     val isFinished: Boolean = false,
 
     @ColumnInfo
-    val tournamentId: Int,
+    val tournamentId: Long,
 )
